+++ Source/Siddon_projection.cu
@@ -105,9 +105,9 @@ __global__ void kernelPixelDetector( Geometry geo,
     
 //     size_t idx = threadIdx.x + blockIdx.x * blockDim.x;
 
-    int y = blockIdx.y * blockDim.y + threadIdx.y;
-    int x = blockIdx.x * blockDim.x + threadIdx.x;
-    size_t idx =  x  * geo.nDetecV + y;
+    unsigned long y = blockIdx.y * blockDim.y + threadIdx.y;
+    unsigned long x = blockIdx.x * blockDim.x + threadIdx.x;
+    unsigned long idx =  x  * geo.nDetecV + y;
 
     if ((x>= geo.nDetecU) | (y>= geo.nDetecV))
         return;
@@ -210,8 +210,8 @@ __global__ void kernelPixelDetector( Geometry geo,
     float sum=0;
     int Np=(imax-imin+1)+(jmax-jmin+1)+(kmax-kmin+1); // Number of intersections
     // Go iterating over the line, intersection by intersection. If double point, no worries, 0 will be computed
-
-    for (int ii=0;ii<Np;ii++){
+    
+    for (unsigned int ii=0;ii<Np;ii++){ 
         if (ax==aminc){
             sum+=(ax-ac)*tex3D(tex, i+0.5, j+0.5, k+0.5);
             i=i+iu;
@@ -297,7 +297,7 @@ int siddon_ray_projection(float const * const img, Geometry geo, float** result,
     divV=16;
     dim3 grid((geo.nDetecU+divU-1)/divU,(geo.nDetecV+divV-1)/divV,1);
     dim3 block(divU,divV,1); 
-    for (int i=0;i<nalpha;i++){
+    for (unsigned int i=0;i<nalpha;i++){
         
         geo.alpha=alphas[i];
